+++ src/components/NewsCard/NewsCard.css
@@ -3,11 +3,13 @@
   height: 576px;
   position: relative;
   background: white;
+  border-radius: 15px;
 }
 
 .card__image {
   width: 400px;
   height: 272px;
+  object-fit: cover;
 }
 
 .card__date {
@@ -92,10 +94,6 @@
   background-image: url("../../images/bookmarkBlue.svg");
 }
 
-.card__bookmark-active {
-  background-image: url("../../images/bookmarked.svg");
-}
-
 .card__signin {
   position: absolute;
   top: 24px;
@@ -166,7 +164,7 @@
     margin: 4px 0 4px 16px;
   }
   .card__title {
-    height: 48px;
+    height: 49px;
     width: 256px;
     margin: 8px 0 8px 16px;
     font-size: 22px;
@@ -216,7 +214,7 @@
     margin: 16px 0 12px 18px;
   }
   .card__title {
-    height: 72px;
+    height: 73px;
     width: 192px;
     margin: 12px 0 10px 16px;
     font-size: 22px;
@@ -266,7 +264,7 @@
   }
 }
 
-@media screen and (max-width: 800px) {
+@media screen and (max-width: 767px) {
   .card {
     width: 188px;
     height: 270px;
@@ -307,7 +305,7 @@
   }
 }
 
-@media screen and (max-width: 693px) {
+@media screen and (max-width: 682px) {
   .card {
     width: 288px;
     height: 440px;
@@ -326,8 +324,8 @@
     width: 256px;
     font-size: 22px;
     line-height: 24px;
-    margin: 0 auto 14px 16px;
-    height: 45px;
+    margin: 0 auto 10px 16px;
+    height: 51px;
     -webkit-line-clamp: 2;
   }
   .card__paragraph {
