+++ src/components/SearchForm/SearchForm.css
@@ -1,0 +1,3 @@
+.search-form {
+  position: relative;
+}
