+++ src/index.css
@@ -10,6 +10,7 @@
 @import url("./components/PopupWithForm/PopupWithForm.css");
 @import url("./components/Preloader/Preloader.css");
 @import url("./components/RegisterModal/RegisterModal.css");
+@import url("./components/SearchForm/SearchForm.css");
 @import url("./components/SearchResultsHome/SearchResultsHome.css");
 @import url("./components/SignInModal/SignInModal.css");
 @import url("./components/SuccessfulModal/SuccessfulModal.css");
